rhdl

Module prelude

Re-exports

pub use rhdl_trace_type as rtt;

pub use rhdl_vlog as vlog;

Modules

drc

reset

trace

Macros

bind

const_max

export

op

parse_quote_miette

path

Structs

Adapter

BitZ

Bits

The Bits type is a fixed-sized bit vector. It is meant to imitate the behavior of bit vectors in hardware. Due to the design of the Bits type, you can only create a Bits type of up to 128 bits in length for now. However, you can easily express larger constructs in hardware using arrays, tuples and structs. The only real limitation of the Bits type being 128 bits is that you cannot perform arbitrary arithmetic on longer bit values in your hardware designs. I don’t think this is a significant issue, but the Bits design of the rust-hdl crate was much slower and harder to maintain and use. I think this is a good trade-off.

Blue

CircuitDescriptor

Clock

ClockReset

Const

Driver

Fixture

Func

Green

HDLDescriptor

Indigo

NoKernel2

NoKernel3

Object

Orange

Path

Red

Reset

ResetN

Signal

SignedBits

The SignedBits type is a fixed-size bit vector. It is meant to imitate the behavior of signed bit vectors in hardware. Due to the design of the SignedBits type, you can only create a signed bit vector of up to 128 bits in lnegth for now. However, you can easily express larger constructs in hardware using arrays, tuples and structs. The only real limitation of the SignedBits type being 128 bits is that you cannot perform arbitrary arithmetic on longer bit values in your hardware designs.

SvgOptions

SynchronousTestBench

TestBench

TestBenchOptions

TimedSample

U1

U2

U3

U4

U5

U6

U7

U8

U9

U10

U11

U12

U13

U14

U15

U16

U17

U18

U19

U20

U21

U22

U23

U24

U25

U26

U27

U28

U29

U30

U31

U32

U33

U34

U35

U36

U37

U38

U39

U40

U41

U42

U43

U44

U45

U46

U47

U48

U49

U50

U51

U52

U53

U54

U55

U56

U57

U58

U59

U60

U61

U62

U63

U64

U65

U66

U67

U68

U69

U70

U71

U72

U73

U74

U75

U76

U77

U78

U79

U80

U81

U82

U83

U84

U85

U86

U87

U88

U89

U90

U91

U92

U93

U94

U95

U96

U97

U98

U99

U100

U101

U102

U103

U104

U105

U106

U107

U108

U109

U110

U111

U112

U113

U114

U115

U116

U117

U118

U119

U120

U121

U122

U123

U124

U125

U126

U127

U128

Vcd

Violet

Yellow

bits

signed

Enums

BitString

BitX

CompilationMode

ExportError

Kind

MountPoint

OpCode

RHDLError

Traits

BitWidth

Circuit

CircuitDQ

CircuitIO

ClockPosEdgeExt

Digital

This is the core trait for all of RHDL data elements. If you want to use a data type in the hardware part of the design, it must implement this trait.

DigitalFn

DigitalFn0

DigitalFn1

DigitalFn2

DigitalFn3

DigitalFn4

DigitalFn5

DigitalFn6

Domain

IsEqualTo

IsGreaterThan

IsGreaterThanOrEqualTo

IsLessThan

IsLessThanOrEqualTo

Max

MergeExt

Pretty

ProbeExt

RunExt

RunSynchronousExt

RunSynchronousFeedbackExt

RunWithoutSynthesisSynchronousExt

Synchronous

SynchronousDQ

SynchronousIO

SynchronousProbeExt

Timed

TimedStreamExt

TraceKey

Unsigned

XAdd

XMul

XNeg

XSgn

XSub

Functions

b1

b2

b3

b4

b5

b6

b7

b8

b9

b10

b11

b12

b13

b14

b15

b16

b17

b18

b19

b20

b21

b22

b23

b24

b25

b26

b27

b28

b29

b30

b31

b32

b33

b34

b35

b36

b37

b38

b39

b40

b41

b42

b43

b44

b45

b46

b47

b48

b49

b50

b51

b52

b53

b54

b55

b56

b57

b58

b59

b60

b61

b62

b63

b64

b65

b66

b67

b68

b69

b70

b71

b72

b73

b74

b75

b76

b77

b78

b79

b80

b81

b82

b83

b84

b85

b86

b87

b88

b89

b90

b91

b92

b93

b94

b95

b96

b97

b98

b99

b100

b101

b102

b103

b104

b105

b106

b107

b108

b109

b110

b111

b112

b113

b114

b115

b116

b117

b118

b119

b120

b121

b122

b123

b124

b125

b126

b127

b128

bit_range

bits

Helper function for creating a bits value from a constant.

bitx_parse

bitx_string

bitx_vec

circuit_black_box

clock

clock_reset

compile_design

compile_design_stage1

constant

execute

hash_id

merge

passthrough_input_driver

passthrough_output_driver

reset

reset_n

run_async_red_blue

s2

s3

s4

s5

s6

s7

s8

s9

s10

s11

s12

s13

s14

s15

s16

s17

s18

s19

s20

s21

s22

s23

s24

s25

s26

s27

s28

s29

s30

s31

s32

s33

s34

s35

s36

s37

s38

s39

s40

s41

s42

s43

s44

s45

s46

s47

s48

s49

s50

s51

s52

s53

s54

s55

s56

s57

s58

s59

s60

s61

s62

s63

s64

s65

s66

s67

s68

s69

s70

s71

s72

s73

s74

s75

s76

s77

s78

s79

s80

s81

s82

s83

s84

s85

s86

s87

s88

s89

s90

s91

s92

s93

s94

s95

s96

s97

s98

s99

s100

s101

s102

s103

s104

s105

s106

s107

s108

s109

s110

s111

s112

s113

s114

s115

s116

s117

s118

s119

s120

s121

s122

s123

s124

s125

s126

s127

s128

signal

signed

Helper function for creating a signed bits value from a constant.

sub_trace_type

synchronous_black_box

timed_sample

trace

trace_init_db

trace_pop_path

trace_push_path

trace_time

with_trace_db

without_reset

Type Aliases

Add1

Maximum

Sum

b1

b2

b3

b4

b5

b6

b7

b8

b9

b10

b11

b12

b13

b14

b15

b16

b17

b18

b19

b20

b21

b22

b23

b24

b25

b26

b27

b28

b29

b30

b31

b32

b33

b34

b35

b36

b37

b38

b39

b40

b41

b42

b43

b44

b45

b46

b47

b48

b49

b50

b51

b52

b53

b54

b55

b56

b57

b58

b59

b60

b61

b62

b63

b64

b65

b66

b67

b68

b69

b70

b71

b72

b73

b74

b75

b76

b77

b78

b79

b80

b81

b82

b83

b84

b85

b86

b87

b88

b89

b90

b91

b92

b93

b94

b95

b96

b97

b98

b99

b100

b101

b102

b103

b104

b105

b106

b107

b108

b109

b110

b111

b112

b113

b114

b115

b116

b117

b118

b119

b120

b121

b122

b123

b124

b125

b126

b127

b128

s2

s3

s4

s5

s6

s7

s8

s9

s10

s11

s12

s13

s14

s15

s16

s17

s18

s19

s20

s21

s22

s23

s24

s25

s26

s27

s28

s29

s30

s31

s32

s33

s34

s35

s36

s37

s38

s39

s40

s41

s42

s43

s44

s45

s46

s47

s48

s49

s50

s51

s52

s53

s54

s55

s56

s57

s58

s59

s60

s61

s62

s63

s64

s65

s66

s67

s68

s69

s70

s71

s72

s73

s74

s75

s76

s77

s78

s79

s80

s81

s82

s83

s84

s85

s86

s87

s88

s89

s90

s91

s92

s93

s94

s95

s96

s97

s98

s99

s100

s101

s102

s103

s104

s105

s106

s107

s108

s109

s110

s111

s112

s113

s114

s115

s116

s117

s118

s119

s120

s121

s122

s123

s124

s125

s126

s127

s128

Attribute Macros

kernel

Derive Macros

Circuit

CircuitDQ

Digital

Synchronous

SynchronousDQ

Timed