rhdl::prelude

Struct SignedBits

pub struct SignedBits<Len>
where
    Len: BitWidth,
{ /* private fields */ }

The SignedBits type is a fixed-size bit vector. It is meant to imitate the behavior of signed bit vectors in hardware. Due to the design of the SignedBits type, you can only create a signed bit vector of up to 128 bits in lnegth for now. However, you can easily express larger constructs in hardware using arrays, tuples and structs. The only real limitation of the SignedBits type being 128 bits is that you cannot perform arbitrary arithmetic on longer bit values in your hardware designs.

Signed arithmetic is performed using 2’s complement arithmetic. See https://en.wikipedia.org/wiki/Two%27s_complement for more information.

Note that unlike the Bits type, comparisons are performed using signed arithmetic. Note also that the right shift operator when applied to a signed value will sign extend the value. This is the same behavior as is seen in Rust (i.e., ((-4) >> 1) == -2).

If you want to right shift a signed value without sign extension, then you should convert it to a Bits type first.

Implementations

impl<N> SignedBits<N>
where N: BitWidth,

pub const MAX: SignedBits<N>

pub const MIN: SignedBits<N>

pub const ZERO: SignedBits<N>

Return a SignedBits value with all bits set to 1.

pub const fn max_value() -> i128

Return the largest positive value that can be represented by this sized SignedBits value.

assert_eq!(SignedBits::<U8>::max_value(), i8::MAX as i128);

pub const fn min_value() -> i128

Return the smallest negative value that can be represented by this sized SignedBits value.

assert_eq!(SignedBits::<U8>::min_value(), i8::MIN as i128);

pub const fn is_negative(&self) -> bool

Test if the value is negative.

assert!(signed::<U8>(-1).is_negative());
assert!(!signed::<U8>(0).is_negative());
assert!(!signed::<U8>(1).is_negative());

pub const fn is_non_negative(&self) -> bool

Test if the value is positive or zero.

assert!(!signed::<U8>(-1).is_non_negative());
assert!(signed::<U8>(0).is_non_negative());
assert!(signed::<U8>(1).is_non_negative());

pub const fn as_unsigned(self) -> Bits<N>

Reinterpret the SignedBits value as an unsigned Bits value. This is useful for performing bit manipulations on the value that may or not preserve the 2’s complement nature of the value.

let x = signed::<U8>(-14); // In binary: 1111_0010
let y : Bits<U8> = x.as_unsigned();
assert_eq!(y, 0b1111_0010);

pub const fn raw(self) -> i128

Extract the raw signed i128 backing this SignedBits value.

pub const fn dyn_bits(self) -> SignedDynBits

Convert the compile time sized SignedBits value to a run-time traced [SignedDynBits] value.

pub fn to_bools(self) -> Vec<bool>

Build a (dynamic, stack allocated) vector containing the bits that make up this value. This will be slow.

pub const fn any(self) -> bool

pub const fn all(self) -> bool

pub const fn xor(self) -> bool

pub const fn xext<M>(self) -> SignedBits<<N as Add<M>>::Output>
where N: Add<M>, <N as Add<M>>::Output: BitWidth,

Sign extend a value by the number of bits specified.

pub const fn xshr<M>(self) -> SignedBits<<N as Sub<M>>::Output>
where N: Sub<M>, M: BitWidth, <N as Sub<M>>::Output: BitWidth,

pub const fn xshl<M>(self) -> SignedBits<<N as Add<M>>::Output>
where N: Add<M>, M: BitWidth, <N as Add<M>>::Output: BitWidth,

pub const fn resize<M>(self) -> SignedBits<M>
where M: BitWidth,