rhdl::prelude

Trait Digital

pub trait Digital:
    Sized
    + Copy
    + PartialEq
    + Clone
    + 'static {
    const BITS: usize;
    const TRACE_BITS: usize = Self::BITS;

    // Required methods
    fn static_kind() -> Kind;
    fn static_trace_type() -> TraceType;
    fn bin(self) -> Vec<BitX>;
    fn dont_care() -> Self;

    // Provided methods
    fn bits() -> usize { ... }
    fn trace_bits() -> usize { ... }
    fn kind(&self) -> Kind { ... }
    fn trace_type(&self) -> TraceType { ... }
    fn trace(self) -> Vec<TraceBit> { ... }
    fn typed_bits(self) -> TypedBits { ... }
    fn discriminant(self) -> TypedBits { ... }
    fn variant_kind(self) -> Kind { ... }
    fn binary_string(self) -> String { ... }

}

This is the core trait for all of RHDL data elements. If you want to use a data type in the hardware part of the design, it must implement this trait.

From https://serde.rs/data-model.html, we get a catalog of different types that represent the data model used by Serde (and by extension Rust). Here are the list of types and how they are represented in RHDL.

Primitive types

Only the bool type is directly supported. Otherwise, use Bits or SignedBits to ensure that the arithmetic operations model the behavior of the hardware.

String, Byte Array, Unit, Unit Struct, Sequence, Map

These are all unsupported on a hardware target. They either have variable size or no size at all.

Option

The option is supported in RHDL. It is represented as an enum with two variants, precisely as it is in Rust.

Enum Variants

Enum variants can be either empty or have a payload. Empty variants are represented via the discriminant, with no payload. Variants with a payload are represented as a the discriminant and the payload packed into binary representation.

Structs, Tuples, Arrays, Unions

These are all supported in RHDL.

Required Associated Constants

const BITS: usize

Provided Associated Constants

const TRACE_BITS: usize = Self::BITS

Required Methods

fn static_kind() -> Kind

fn static_trace_type() -> TraceType

fn bin(self) -> Vec<BitX>

fn dont_care() -> Self

Provided Methods

fn bits() -> usize

fn trace_bits() -> usize

fn kind(&self) -> Kind

fn trace_type(&self) -> TraceType

fn trace(self) -> Vec<TraceBit>

fn typed_bits(self) -> TypedBits

fn discriminant(self) -> TypedBits

fn variant_kind(self) -> Kind

fn binary_string(self) -> String

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.